Widen TextOverlay text-element handles to uintptr for x64 (x64 fix)
CTextOverlay's on-screen-text API used a uint32 as an STextElement*
handle: u4DisplayFormattedString / u4DisplayPositionedString return
(uint32)ptel and take a uint32 u4_prev to chain sequences, and
u4FindSequenceEnd returns (uint32)&elem. On x64 the element address was
truncated to 32 bits -> reconstructing (STextElement*)u4_prev gives a
garbage pointer -> crash when chaining subtitles/tutorial text. Widen
the three functions' return type and the u4_prev params to uintptr (+
the (uintptr)ptel casts), and the two callers' u4_prev locals
(AudioDaemon subtitles, Action tutorial text). uintptr==uint32 on Win32
so behaviour is unchanged there (Win32 green); x64 warning-free.

+++ jp2_pc/Source/Lib/Audio/AudioDaemon.cpp
@@ -2880,7 +2880,7 @@ void CAudioDaemon::ProcessSubtitle
 	//
 	uint8* pu1_data = (uint8*)(psth+1);
 
-	uint32	u4_prev = CTextOverlay::ptovTextSystem->u4FindSequenceEnd(ettSUBTITLE);
+	uintptr	u4_prev = CTextOverlay::ptovTextSystem->u4FindSequenceEnd(ettSUBTITLE);
 	TSec	s_time;
 
 	// go through all of the sections in the data block...
